Now it works with INT_MAX and INT_MIN

+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: alvera-v <alvera-v@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/25 11:05:05 by alvera-v          #+#    #+#             */
-/*   Updated: 2025/05/27 12:30:26 by alvera-v         ###   ########.fr       */
+/*   Updated: 2025/05/28 12:34:56 by alvera-v         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ void				print_error(t_Stack **stack, char **argv,
 void				free_string(char **string);
 void				free_split(char **split);
 void				start_sorting(t_Stack *stack_a, t_Stack *stack_b);
-int					ft_atoi(const char *nptr);
+long				ft_atoi(const char *nptr);
 int					ft_isdigit(char c);
 char				**ft_split(char *s, char c);
 size_t				ft_strlen(char const *s);
